🚚 rename: fixTypo 및 변경에 따른 import문 수정  (Jira Ticket : CCC-66)
- oauth/_component -> oauth/_components (fixTypo)
- 불필요한 react import 문 제거
- utils, schema, error-boundary, api constants 변경에 따른 import문 수정
- test 페이지 제거
- user-history -> (user)/user-history (폴더이동)
