+++ app/(user)/user-history/_components/HistoryItem.tsx
@@ -1,7 +1,6 @@
 import CheckIcon from '@/public/icons/user-history/check_icon.svg';
 import HamburgerIcon from '@/public/icons/user-history/hamburger_icon.svg';
 import { Task } from '@ccc-types';
-import React from 'react';
 
 const HistoryItem = ({ task }: { task: Task }) => (
   <div className="flex items-center gap-1 rounded-lg bg-customBackground-secondary px-[14px] py-[10px]">
