+++ app/[teamId]/_components/TeamToDoListCard.tsx
@@ -5,9 +5,9 @@ import {
   ChartTooltip,
   ChartTooltipContent,
 } from '@/components/ui/chart';
+import { stringToHex } from '@/lib/utils';
 import Kebab from '@/public/icons/kebab.svg';
 import ToDoDoneIcon from '@/public/icons/todo_done.svg';
-import stringToHex from '@/utils/StringToColor';
 import { useMemo } from 'react';
 import { Pie, PieChart } from 'recharts';
