+++ app/[teamId]/tasks/_components/CommentSheet/CommentMeta.tsx
@@ -1,9 +1,8 @@
+import { dateFormatter } from '@/lib/utils';
 import CalenderNoBtnIcon from '@/public/icons/list/calender_no_btn.svg';
 import ClockIcon from '@/public/icons/list/clock_icon.svg';
 import DailyIcon from '@/public/icons/list/daily_task_icon.svg';
-import { formatToDate, formatToTime } from '@/utils/dateFormat';
 import Image from 'next/image';
-import React from 'react';
 
 const textClass = `text-xs font-normal text-text-default`;
 
@@ -23,19 +22,19 @@ function CommentMeta() {
           <span className="text-sm font-medium">안혜나</span>
         </div>
         <span className="text-sm font-normal text-text-secondary">
-          {formatToDate(new Date(), 'dotFormat')}
+          {dateFormatter.toConvertDate(new Date(), 'dotFormat')}
         </span>
       </div>
       <div className="flex gap-3">
         <div className="flex items-center gap-1">
           <CalenderNoBtnIcon />
           <p className={textClass}>
-            {formatToDate(new Date(), 'koreanFullDate')}
+            {dateFormatter.toConvertDate(new Date(), 'koreanFullDate')}
           </p>
         </div>
         <div className="flex items-center gap-1">
           <ClockIcon />
-          <p className={textClass}>{formatToTime(new Date())}</p>
+          <p className={textClass}>{dateFormatter.toTime(new Date())}</p>
         </div>
         <div className="flex items-center gap-1">
           <DailyIcon />
