--- app/test/page.tsx
@@ -1,12 +1,0 @@
-'use client';
-
-import UserList from './_components/UserList';
-
-function page() {
-  return (
-    <div>
-      <UserList />
-    </div>
-  );
-}
-export default page;
