+++ components/modal-template/LogoutModal.tsx
@@ -7,7 +7,6 @@ import {
   DialogTitle,
   DialogTrigger,
 } from '@/components/ui/dialog';
-import React from 'react';
 
 function LogoutModal({ className = '', ...props }) {
   return (
