+++ components/modal-template/WithDrawalModal.tsx
@@ -11,9 +11,8 @@ import {
 } from '@/components/ui/dialog';
 import WarningIcon from '@/public/icons/modal/warning_icon.svg';
 import WithDrawIcon from '@/public/icons/modal/with_draw_icon.svg';
-import React from 'react';
 
-function WithDrawalTemplete({ className = '', ...props }) {
+function WithDrawalTemplate({ className = '', ...props }) {
   return (
     <Dialog>
       <DialogTrigger asChild className={className}>
@@ -47,4 +46,4 @@ function WithDrawalTemplete({ className = '', ...props }) {
   );
 }
 
-export default WithDrawalTemplete;
+export default WithDrawalTemplate;
