+++ components/ui/PasswordInput.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import cn from '@/lib/utils';
+import { cn } from '@/lib/utils';
 import HidePasswordIcon from '@/public/icons/visibility_off.svg';
 import ShowPasswordIcon from '@/public/icons/visibility_on.svg';
 import * as React from 'react';
