+++ components/ui/checkbox.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import cn from '@/lib/utils';
+import { cn } from '@/lib/utils';
 import * as CheckboxPrimitive from '@radix-ui/react-checkbox';
 import { Check } from 'lucide-react';
 import * as React from 'react';
