+++ components/ui/dialog.tsx
@@ -1,4 +1,4 @@
-import cn from '@/lib/utils';
+import { cn } from '@/lib/utils';
 import CloasIcon from '@/public/icons/modal/close_icon.svg';
 // eslint-disable-next-line import/no-extraneous-dependencies
 import * as DialogPrimitive from '@radix-ui/react-dialog';
