+++ components/ui/form.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import { Label } from '@/components/ui/label';
-import cn from '@/lib/utils';
+import { cn } from '@/lib/utils';
 import * as LabelPrimitive from '@radix-ui/react-label';
 import { Slot } from '@radix-ui/react-slot';
 import * as React from 'react';
