+++ components/ui/label.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import cn from '@/lib/utils';
+import { cn } from '@/lib/utils';
 import * as LabelPrimitive from '@radix-ui/react-label';
 import { type VariantProps, cva } from 'class-variance-authority';
 import * as React from 'react';
