+++ lib/api/common.ts
@@ -1,6 +1,6 @@
 'use client';
 
-import { ENDPOINTS } from '@/lib/api/API_CONSTANTS';
+import ENDPOINTS from '@/lib/api/ENDPOINTS';
 import client from '@/lib/api/client/client';
 import { UrlType } from '@ccc-types';
