+++ lib/api/oAuth.ts
@@ -1,4 +1,4 @@
-import { ENDPOINTS } from '@/lib/api/API_CONSTANTS';
+import ENDPOINTS from '@/lib/api/ENDPOINTS';
 import client from '@/lib/api/client/client';
 import { AppSecret, OAuthApp, OAuthProvider, UrlType } from '@ccc-types';
